Extract module CodeLength.
