+++ lib/rubocop.rb
@@ -67,6 +67,7 @@ require 'rubocop/cop/style/case_equality'
 require 'rubocop/cop/style/case_indentation'
 require 'rubocop/cop/style/character_literal'
 require 'rubocop/cop/style/class_and_module_camel_case'
+require 'rubocop/cop/style/code_length'
 require 'rubocop/cop/style/class_length'
 require 'rubocop/cop/style/class_methods'
 require 'rubocop/cop/style/class_vars'
